Adapterをクラス化
