+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".fragment.SearchActivity"></activity>
+        <activity android:name=".activity.SearchActivity"></activity>
         <activity android:name=".activity.WebViewActivity" />
         <activity android:name=".activity.MainActivity">
             <intent-filter>
